Crate general

Contains some general helper functions that are not specific to one area but might be used in different parts of the project

Structs

Name

A single Name for a Resource, which consists of a user given Name and a Groupt to distinguish between two Resources that may have the same Name but come from two different Systems/Sources

Shared

A Datastructure that allows Data to be shared across threads/tasks and be updated, by replacing it atomically.

Enums

Group

The Group to which a Resource belongs or was loaded from, like Kubernetes or a File

Functions

parse_time

Parses a simple String representation for Time into a Duration struct to allow for easier usage